requirements.txt
requests
rsa
attrs
aiohttp
toml
Pillow
argon2-cffi